graph: Move EntityOp to EntityCache
That's the only placed where it is needed
